# api/openapi.yaml
components:
  schemas:
...
    ChannelResponse:
      properties:
        authData:
          $ref: '#/components/schemas/AuthData'
          description: WEB channels only (docs/05-api.md Abschnitt 3) - never present for APP.
        channel:
          $ref: '#/components/schemas/ChannelBlock'
        demo:
          $ref: '#/components/schemas/DemoInfo'
          description: Demo-only correlation IDs, never part of the production contract.
        next:
          $ref: '#/components/schemas/Next'
        stepData:
          $ref: '#/components/schemas/StepData'
          description: Whatever the current step needs to render. `kind` names the shape - see StepData.
      required:
      - channel
      type: object
...
    ConfirmQrLoginActivateRequest:
      properties:
        pairingCode:
          example: ABCD-1234
          type: string
      type: object
...
    Next:
      properties:
        context:
          example: authentication
          type: string
        step:
          example: auth
          type: string
        toolId:
          example: auth-sms
          type: string
        toolSessionId:
          format: uuid
          type: string
        type:
          example: tool
          type: string
      required:
      - step
      - type
      type: object
...
  securitySchemes:
    dpop:
      description: DPoP proof JWT, one per request, bound to method and URL (docs/09-dpop.md). The App channel's proof; the
        binding key is derived from it and is never supplied by the caller.
      in: header
      name: DPoP
      type: apiKey
    kc-peer-auth:
      bearerFormat: JWT
      description: 'Signed peer-auth assertion (docs/05-api.md Abschnitt 3, ADR-7). The Web channel''s proof: Keycloak calls
        server-to-server, and the channel binding comes out of the assertion.'
      scheme: bearer
      type: http
info:
  description: App-facing (orchestrator-first) API for the DPoP-bound registration and login demo. See docs/05-api.md for
    the full contract; the Keycloak-facing kc facade is out of scope for this build (docs/08-projektrahmen.md).
  title: Identity-Demo Orchestrator API
  version: v1
openapi: 3.1.0
paths:
...
  /orchestrator/api/v1/channels/{channelSessionId}/tools/auth-device:
    post:
      description: 'No request body: toolId already carries kind and method.'
      operationId: activate_22
      parameters:
      - in: path
        name: channelSessionId
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '201':
          content:
            application/json:
              example:
                channel:
                  channelSessionId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  state: ANONYMOUS
                next:
                  step: auth
                  toolId: auth-device
                  toolSessionId: 9c858901-8a57-4791-81fe-4c455b099bc9
                  type: tool
              schema:
                $ref: '#/components/schemas/ChannelResponse'
          description: Created
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Error. `error` names the case; its HTTP status is fixed per code (see ErrorResponse).
      security:
      - dpop: [
          ]
      - kc-peer-auth: [
          ]
      summary: Activate auth-device
      tags:
      - 'Tool: Gerät'
  /orchestrator/api/v1/channels/{channelSessionId}/tools/auth-email:
    post:
      description: 'No request body: toolId already carries kind and method.'
      operationId: activate_20
      parameters:
      - in: path
        name: channelSessionId
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '201':
          content:
            application/json:
              example:
                channel:
                  channelSessionId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  currentAcr: loa1
                  state: STEP_UP_IN_PROGRESS
                next:
                  step: auth
                  toolId: auth-email
                  toolSessionId: 9c858901-8a57-4791-81fe-4c455b099bc9
                  type: tool
              schema:
                $ref: '#/components/schemas/ChannelResponse'
          description: Created
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Error. `error` names the case; its HTTP status is fixed per code (see ErrorResponse).
      security:
      - dpop: [
          ]
      - kc-peer-auth: [
          ]
      summary: Activate auth-email
      tags:
      - 'Tool: E-Mail'
  /orchestrator/api/v1/channels/{channelSessionId}/tools/auth-email-lookup:
    post:
      description: 'No request body: toolId already carries kind and method.'
      operationId: activate_21
      parameters:
      - in: path
        name: channelSessionId
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '201':
          content:
            application/json:
              example:
                channel:
                  channelSessionId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  state: REGISTERING
                next:
                  step: auth
                  toolId: auth-email-lookup
                  toolSessionId: 9c858901-8a57-4791-81fe-4c455b099bc9
                  type: tool
              schema:
                $ref: '#/components/schemas/ChannelResponse'
          description: Created
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Error. `error` names the case; its HTTP status is fixed per code (see ErrorResponse).
      security:
      - dpop: [
          ]
      - kc-peer-auth: [
          ]
      summary: Activate auth-email-lookup
      tags:
      - 'Tool: E-Mail'
  /orchestrator/api/v1/channels/{channelSessionId}/tools/auth-invite:
    post:
      description: 'No request body: toolId already carries kind and method.'
      operationId: activate_19
      parameters:
      - in: path
        name: channelSessionId
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '201':
          content:
            application/json:
              example:
                channel:
                  channelSessionId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  state: ANONYMOUS
                next:
                  step: input
                  toolId: auth-invite
                  toolSessionId: 9c858901-8a57-4791-81fe-4c455b099bc9
                  type: tool
              schema:
                $ref: '#/components/schemas/ChannelResponse'
          description: Created
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Error. `error` names the case; its HTTP status is fixed per code (see ErrorResponse).
      security:
      - dpop: [
          ]
      - kc-peer-auth: [
          ]
      summary: Activate auth-invite
      tags:
      - 'Tool: Einmalkennwort'
  /orchestrator/api/v1/channels/{channelSessionId}/tools/auth-kobil:
    post:
      operationId: activate_18
      parameters:
      - in: path
        name: channelSessionId
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '201':
          content:
            application/json:
              example:
                channel:
                  channelSessionId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  state: ANONYMOUS
                next:
                  step: unlock
                  toolId: auth-kobil
                  toolSessionId: 9c858901-8a57-4791-81fe-4c455b099bc9
                  type: tool
                stepData:
                  kind: kobil-unlock
                  kobilUserId: kob-1a2b3c4d5e6f
                  tenantId: identity-demo
                  unlockOptions:
                  - biometric
                  - password
              schema:
                $ref: '#/components/schemas/ChannelResponse'
          description: Created
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Error. `error` names the case; its HTTP status is fixed per code (see ErrorResponse).
      security:
      - dpop: [
          ]
      - kc-peer-auth: [
          ]
      summary: Activate auth-kobil
      tags:
      - 'Tool: KOBIL'
  /orchestrator/api/v1/channels/{channelSessionId}/tools/auth-password:
    post:
      description: 'No request body: toolId already carries kind and method.'
      operationId: activate_16
      parameters:
      - in: path
        name: channelSessionId
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '201':
          content:
            application/json:
              example:
                channel:
                  channelSessionId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  currentAcr: loa1
                  state: STEP_UP_IN_PROGRESS
                next:
                  step: auth
                  toolId: auth-password
                  toolSessionId: 9c858901-8a57-4791-81fe-4c455b099bc9
                  type: tool
              schema:
                $ref: '#/components/schemas/ChannelResponse'
          description: Created
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Error. `error` names the case; its HTTP status is fixed per code (see ErrorResponse).
      security:
      - dpop: [
          ]
      - kc-peer-auth: [
          ]
      summary: Activate auth-password
      tags:
      - 'Tool: Passwort'
  /orchestrator/api/v1/channels/{channelSessionId}/tools/auth-password-lookup:
    post:
      description: 'No request body: toolId already carries kind and method.'
      operationId: activate_17
      parameters:
      - in: path
        name: channelSessionId
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '201':
          content:
            application/json:
              example:
                channel:
                  channelSessionId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  state: REGISTERING
                next:
                  step: auth
                  toolId: auth-password-lookup
                  toolSessionId: 9c858901-8a57-4791-81fe-4c455b099bc9
                  type: tool
              schema:
                $ref: '#/components/schemas/ChannelResponse'
          description: Created
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Error. `error` names the case; its HTTP status is fixed per code (see ErrorResponse).
      security:
      - dpop: [
          ]
      - kc-peer-auth: [
          ]
      summary: Activate auth-password-lookup
      tags:
      - 'Tool: Passwort'
  /orchestrator/api/v1/channels/{channelSessionId}/tools/auth-qr:
    post:
      description: 'No request body: toolId already carries kind and method.'
      operationId: activate_14
      parameters:
      - in: path
        name: channelSessionId
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '200':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ChannelResponse'
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Error. `error` names the case; its HTTP status is fixed per code (see ErrorResponse).
      security:
      - dpop: [
          ]
      - kc-peer-auth: [
          ]
      summary: Activate auth-qr
      tags:
      - 'Tool: QR-Login'
  /orchestrator/api/v1/channels/{channelSessionId}/tools/auth-qr-lookup:
    post:
      description: 'No request body: toolId already carries kind and method.'
      operationId: activate_15
      parameters:
      - in: path
        name: channelSessionId
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '200':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ChannelResponse'
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Error. `error` names the case; its HTTP status is fixed per code (see ErrorResponse).
      security:
      - dpop: [
          ]
      - kc-peer-auth: [
          ]
      summary: Activate auth-qr-lookup
      tags:
      - 'Tool: QR-Login'
  /orchestrator/api/v1/channels/{channelSessionId}/tools/auth-sms:
    post:
      description: 'No request body: toolId already carries kind and method.'
      operationId: activate_12
      parameters:
      - in: path
        name: channelSessionId
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '201':
          content:
            application/json:
              example:
                channel:
                  channelSessionId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  currentAcr: loa1
                  state: STEP_UP_IN_PROGRESS
                next:
                  step: auth
                  toolId: auth-sms
                  toolSessionId: 9c858901-8a57-4791-81fe-4c455b099bc9
                  type: tool
              schema:
                $ref: '#/components/schemas/ChannelResponse'
          description: Created
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Error. `error` names the case; its HTTP status is fixed per code (see ErrorResponse).
      security:
      - dpop: [
          ]
      - kc-peer-auth: [
          ]
      summary: Activate auth-sms
      tags:
      - 'Tool: SMS'
  /orchestrator/api/v1/channels/{channelSessionId}/tools/auth-sms-lookup:
    post:
      description: 'No request body: toolId already carries kind and method.'
      operationId: activate_13
      parameters:
      - in: path
        name: channelSessionId
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '201':
          content:
            application/json:
              example:
                channel:
                  channelSessionId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  state: REGISTERING
                next:
                  step: auth
                  toolId: auth-sms-lookup
                  toolSessionId: 9c858901-8a57-4791-81fe-4c455b099bc9
                  type: tool
              schema:
                $ref: '#/components/schemas/ChannelResponse'
          description: Created
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Error. `error` names the case; its HTTP status is fixed per code (see ErrorResponse).
      security:
      - dpop: [
          ]
      - kc-peer-auth: [
          ]
      summary: Activate auth-sms-lookup
      tags:
      - 'Tool: SMS'
  /orchestrator/api/v1/channels/{channelSessionId}/tools/confirm-email:
    post:
      description: 'No request body: toolId already carries kind and method.'
      operationId: activate_11
      parameters:
      - in: path
        name: channelSessionId
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '201':
          content:
            application/json:
              example:
                channel:
                  channelSessionId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  state: REGISTERING
                next:
                  step: input
                  toolId: confirm-email
                  toolSessionId: 9c858901-8a57-4791-81fe-4c455b099bc9
                  type: tool
              schema:
                $ref: '#/components/schemas/ChannelResponse'
          description: Created
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Error. `error` names the case; its HTTP status is fixed per code (see ErrorResponse).
      security:
      - dpop: [
          ]
      - kc-peer-auth: [
          ]
      summary: Activate confirm-email
      tags:
      - 'Tool: E-Mail'
  /orchestrator/api/v1/channels/{channelSessionId}/tools/confirm-qr-login:
    post:
      description: 'Optional body: {pairingCode}, when already known (e.g. from a demo-link deep link) - skips the input step.'
      operationId: activate_10
      parameters:
      - in: path
        name: channelSessionId
        required: true
        schema:
          format: uuid
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmQrLoginActivateRequest'
      responses:
        '200':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ChannelResponse'
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Error. `error` names the case; its HTTP status is fixed per code (see ErrorResponse).
      security:
      - dpop: [
          ]
      - kc-peer-auth: [
          ]
      summary: Activate confirm-qr-login
      tags:
      - 'Tool: QR-Login'
  /orchestrator/api/v1/channels/{channelSessionId}/tools/enroll-device:
    post:
      description: 'No request body: toolId already carries kind and method.'
      operationId: activate_9
      parameters:
      - in: path
        name: channelSessionId
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '201':
          content:
            application/json:
              example:
                channel:
                  channelSessionId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  currentAcr: loa1
                  currentAmr:
                  - password
                  state: AUTHENTICATED
                next:
                  step: enroll
                  toolId: enroll-device
                  toolSessionId: 9c858901-8a57-4791-81fe-4c455b099bc9
                  type: tool
              schema:
                $ref: '#/components/schemas/ChannelResponse'
          description: Created
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Error. `error` names the case; its HTTP status is fixed per code (see ErrorResponse).
      security:
      - dpop: [
          ]
      - kc-peer-auth: [
          ]
      summary: Activate enroll-device
      tags:
      - 'Tool: Gerät'
  /orchestrator/api/v1/channels/{channelSessionId}/tools/enroll-email:
    post:
      description: 'One shot: no request body, and the response already carries the completed outcome.'
      operationId: activate_8
      parameters:
      - in: path
        name: channelSessionId
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '201':
          content:
            application/json:
              example:
                channel:
                  channelSessionId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  state: AUTHENTICATED
                next:
                  context: authentication
                  step: done
                  type: orchestrator
              schema:
                $ref: '#/components/schemas/ChannelResponse'
          description: Created
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Error. `error` names the case; its HTTP status is fixed per code (see ErrorResponse).
      security:
      - dpop: [
          ]
      - kc-peer-auth: [
          ]
      summary: Activate enroll-email
      tags:
      - 'Tool: E-Mail-Login'
  /orchestrator/api/v1/channels/{channelSessionId}/tools/enroll-kobil:
    post:
      description: Provisions the KOBIL user and returns everything the app's SDK needs for its activation call - including
        the PIN, which this backend mints and keeps.
      operationId: activate_7
      parameters:
      - in: path
        name: channelSessionId
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '201':
          content:
            application/json:
              example:
                channel:
                  channelSessionId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  currentAcr: loa2
                  currentAmr:
                  - fsc
                  state: AUTHENTICATED
                next:
                  step: activate
                  toolId: enroll-kobil
                  toolSessionId: 9c858901-8a57-4791-81fe-4c455b099bc9
                  type: tool
                stepData:
                  activationCode: K7M2PQX9
                  kind: kobil-activation
                  kobilUserId: kob-1a2b3c4d5e6f
                  missingFields:
                  - activated
                  - biometricConsent
                  pin: '40318827'
                  tenantId: identity-demo
                  unlockSecret: xE1r...
              schema:
                $ref: '#/components/schemas/ChannelResponse'
          description: Created
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Error. `error` names the case; its HTTP status is fixed per code (see ErrorResponse).
      security:
      - dpop: [
          ]
      - kc-peer-auth: [
          ]
      summary: Activate enroll-kobil
      tags:
      - 'Tool: KOBIL'
  /orchestrator/api/v1/channels/{channelSessionId}/tools/enroll-password:
    post:
      description: 'No request body: toolId already carries kind and method.'
      operationId: activate_6
      parameters:
      - in: path
        name: channelSessionId
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '201':
          content:
            application/json:
              example:
                channel:
                  channelSessionId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  state: REGISTERING
                next:
                  step: enroll
                  toolId: enroll-password
                  toolSessionId: 9c858901-8a57-4791-81fe-4c455b099bc9
                  type: tool
              schema:
                $ref: '#/components/schemas/ChannelResponse'
          description: Created
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Error. `error` names the case; its HTTP status is fixed per code (see ErrorResponse).
      security:
      - dpop: [
          ]
      - kc-peer-auth: [
          ]
      summary: Activate enroll-password
      tags:
      - 'Tool: Passwort'
  /orchestrator/api/v1/channels/{channelSessionId}/tools/enroll-qr:
    post:
      description: 'No request body: toolId already carries kind and method.'
      operationId: activate_5
      parameters:
      - in: path
        name: channelSessionId
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '200':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ChannelResponse'
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Error. `error` names the case; its HTTP status is fixed per code (see ErrorResponse).
      security:
      - dpop: [
          ]
      - kc-peer-auth: [
          ]
      summary: Activate enroll-qr
      tags:
      - 'Tool: QR-Login'
  /orchestrator/api/v1/channels/{channelSessionId}/tools/enroll-sms:
    post:
      description: 'No request body: toolId already carries kind and method.'
      operationId: activate_4
      parameters:
      - in: path
        name: channelSessionId
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '201':
          content:
            application/json:
              example:
                channel:
                  channelSessionId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  state: REGISTERING
                next:
                  step: enroll
                  toolId: enroll-sms
                  toolSessionId: 9c858901-8a57-4791-81fe-4c455b099bc9
                  type: tool
              schema:
                $ref: '#/components/schemas/ChannelResponse'
          description: Created
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Error. `error` names the case; its HTTP status is fixed per code (see ErrorResponse).
      security:
      - dpop: [
          ]
      - kc-peer-auth: [
          ]
      summary: Activate enroll-sms
      tags:
      - 'Tool: SMS'
  /orchestrator/api/v1/channels/{channelSessionId}/tools/ident-eid:
    post:
      description: 'No request body: toolId already carries kind and method.'
      operationId: activate_3
      parameters:
      - in: path
        name: channelSessionId
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '201':
          content:
            application/json:
              example:
                channel:
                  channelSessionId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  state: REGISTERING
                next:
                  step: input
                  toolId: ident-eid
                  toolSessionId: 9c858901-8a57-4791-81fe-4c455b099bc9
                  type: tool
                stepData:
                  kind: missing-fields
                  missingFields:
                  - familyName
                  - givenNames
                  - birthDate
                  - streetAddress
                  - postalCode
                  - locality
                  - restrictedId
              schema:
                $ref: '#/components/schemas/ChannelResponse'
          description: Created
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Error. `error` names the case; its HTTP status is fixed per code (see ErrorResponse).
      security:
      - dpop: [
          ]
      - kc-peer-auth: [
          ]
      summary: Activate ident-eid
      tags:
      - 'Tool: eID'
  /orchestrator/api/v1/channels/{channelSessionId}/tools/ident-fsc:
    post:
      description: 'No request body: toolId already carries kind and method.'
      operationId: activate_2
      parameters:
      - in: path
        name: channelSessionId
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '201':
          content:
            application/json:
              example:
                channel:
                  channelSessionId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  state: REGISTERING
                next:
                  step: input
                  toolId: ident-fsc
                  toolSessionId: 9c858901-8a57-4791-81fe-4c455b099bc9
                  type: tool
              schema:
                $ref: '#/components/schemas/ChannelResponse'
          description: Created
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Error. `error` names the case; its HTTP status is fixed per code (see ErrorResponse).
      security:
      - dpop: [
          ]
      - kc-peer-auth: [
          ]
      summary: Activate ident-fsc
      tags:
      - 'Tool: Freischaltcode'
...